requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
amqp==2.6.1
apache-superset==1.2.0
apispec==3.3.2
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backoff==1.10.0
billiard==3.6.4.0
bleach==3.3.0
Brotli==1.0.9
cachelib==0.1.1
cassandra-driver==3.25.0
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
clickhouse-driver==0.2.1
clickhouse-sqlalchemy==0.1.6
colorama==0.4.4
contextlib2==21.6.0
convertdate==2.3.2
cron-descriptor==1.2.24
croniter==1.0.15
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-Compress==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-Migrate==3.0.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-WTF==0.14.3
geographiclib==1.52
geomet==0.2.1.post1
geopy==2.1.0
graphlib-backport==1.0.3
gunicorn==20.0.4
hazm==0.7.0
holidays==0.10.3
humanize==3.10.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
jellyfish==0.8.2
Jinja2==2.11.3
jsonschema==3.2.0
kafka-python==2.0.2
kombu==4.6.11
korean-lunar-calendar==0.2.1
libwapiti==0.2.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
msgpack==1.0.2
multidict==5.1.0
networkx==2.5.1
nltk==3.3
numpy==1.21.0
packaging==21.0
pandas==1.2.5
parsedatetime==2.6
pathlib2==2.3.6
pgsanity==0.2.9
pkg-resources==0.0.0
polyline==1.4.0
prison==0.1.3
py==1.10.0
pyaes==1.6.1
pyarrow==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyMeeus==0.5.11
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-geohash==0.8.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.7.6
requests==2.25.1
retry==0.9.2
rsa==4.7.2
segtok==1.5.10
selenium==3.141.0
simplejson==3.17.2
six==1.16.0
slackclient==2.5.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.0
tabulate==0.8.9
Telethon==1.22.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
vine==1.3.0
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-JSON==0.3.3
yake==0.4.8
yarl==1.6.3
